backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
opensearch-py==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
email-validator==2.1.0
aiofiles==23.2.1
prometheus-fastapi-instrumentator==6.1.0
pytest==8.2.1
